move mongo testing to connection_test
